0.7.4 fix kgd(#57)
